requirements_freeze.txt
agentipy==2.0.1.post1
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
anchorpy==0.20.1
anchorpy-core==0.2.0
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
backpack_exchange_sdk==1.0.30
base58==2.1.1
based58==0.1.1
bitarray==3.0.0
borsh-construct==0.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
cohere==5.13.12
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cryptography==44.0.1
cytoolz==1.0.1
dataclasses-json==0.6.7
distro==1.9.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.8
fastavro==1.10.0
filelock==3.17.0
flake8==7.1.2
frozenlist==1.5.0
fsspec==2025.2.0
google-auth==2.38.0
greenlet==3.1.1
griffe==1.5.7
groq==0.18.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.0
idna==3.10
jiter==0.8.2
jsonalias==0.1.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.17
langchain-core==0.3.36
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.8
logfire-api==3.6.1
loguru==0.7.3
marshmallow==3.26.1
mccabe==0.7.0
mistralai==1.5.0
more-itertools==8.14.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.63.2
orjson==3.10.15
packaging==24.2
parsimonious==0.10.0
pillow==11.1.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.5.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-ai==0.0.24
pydantic-ai-slim==0.0.24
pydantic-graph==0.0.24
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
pyheck==0.1.5
PyNaCl==1.5.0
pythclient==0.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rlp==4.1.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
solana==0.35.1
solders==0.21.0
SQLAlchemy==2.0.38
starlette==0.45.3
sumtypes==0.1a6
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
toolz==0.11.2
tqdm==4.67.1
tweepy==4.15.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.8.0
websocket-client==1.8.0
websockets==10.4
win32_setctime==1.2.0
yarl==1.18.3
zstandard==0.23.0
